extract common logic from ResolveGenerate rule
